Move the Vault service response calculator to the common folder
